Isolate CPG archive handling into its own class.
